Move StyleSelector to Overview level and delete excess files
