--- client/src/components/Overview/StyleSelector/index.jsx
@@ -1,3 +1,0 @@
-import StyleSelector from './StyleSelector';
-
-export default StyleSelector;
